feat: transform multiple merge keys to one
There exists an issue in yaml.v3 merge of multiple keys which will
result in a panic. Multiple merge keys is not defined in the yaml
specification, but many available parsers allow it.

This new handler can be chained into a visitor to pre-process a
yaml.Node such that it can then be unmarshaled into a struct without
error. See the tests for this handler for a concrete example.

Minimal usage:

	document := &yaml.Node{}
	_ = yaml.Unmarshal([]byte(input), document)
	visitor, _ := yay.NewVisitor(
		yay.NewMultipleToSingleMergeHandler(),
	)
	_ = visitor.Visit(context.TODO(), document)