fix: multiple << keys to list should allow left->right overrides
The previous implementation missed this point from the spec:

> Keys in mapping nodes earlier in the sequence override keys specified in later mapping nodes.

So, where YAML implementations which allow multiple merge keys and
merge this top-down (the later keys overriding earlier keys), the defined
behavior in the specification allows merging of values only if the key is not
already defined previously in the merge list. The previous implementation
included the merge list in the same order as the original merge keys.

This change includes a functional parameter `WithRetainMergeKeyOrder`, which
allows users who are parsing from non-standard or broken frameworks to follow
the same behavior originally implemented in yay.